Migrationfile ammendments for lpw data support
3 migration files will be amended.
In additon do the following for the Procland file:
    // Remove the $Size//
                    $pricePerSize = $price / $size;

+++ database/migrations/2023_02_05_224726_create_statistics_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('statistics', function (Blueprint $table) {
             $table->id();
             $table->integer('property_id');
-            $table->integer('price');
+            $table->decimal('price',14,0);
             $table->timestamps();
         });
     }
